yaw was way
Day 4 update: previous testing showcased contradictory behaviour. The controller somehow formed a positive feedback loop which caused the drone to fix an error by amplifying it.
Turns out the bug was in the part where the PID output is allocated to each motor. Yaw commands were inverted for some reason. All I had to do to fix it was change the + to - and it worked.
The docs didn't have the default motor direction in them as other firmware generally do. Not a good or bad thing, just unexpected.